Move the permissions manager into services
