+++ src/components/services/Services.js
@@ -29,6 +29,7 @@ import { ContactsLoader } from './ContactsLoader.js'
 import EdgeAccountCallbackManager from './EdgeAccountCallbackManager.js'
 import EdgeContextCallbackManager from './EdgeContextCallbackManager.js'
 import EdgeWalletsCallbackManager from './EdgeWalletsCallbackManager.js'
+import { PermissionsManager } from './PermissionsManager.js'
 
 type Props = { context: EdgeContext }
 
@@ -91,6 +92,7 @@ export class Services extends PureComponent<Props> {
           <EdgeContextCallbackManager />
           <EdgeWalletsCallbackManager />
           <ModalProvider />
+          <PermissionsManager />
         </React.Fragment>
       </Provider>
     )
