+++ src/connectors/MainConnector.js
@@ -9,7 +9,7 @@ import { openDrawer } from '../actions/ScenesActions.js'
 import { showReEnableOtpModal } from '../actions/SettingsActions.js'
 import { selectWallet } from '../actions/WalletActions.js'
 import Main from '../components/Main.ui'
-import { requestPermission } from '../modules/PermissionsManager.js'
+import { requestPermission } from '../components/services/PermissionsManager.js'
 import type { Dispatch } from '../types/reduxTypes.js'
 
 const mapStateToProps = () => ({})
